+++ core/chains/evm/chain_set.go
@@ -12,7 +12,9 @@ import (
 	"github.com/smartcontractkit/sqlx"
 	"go.uber.org/multierr"
 
-	evmclient "github.com/smartcontractkit/chainlink/core/chains/evm/client"
+	"github.com/smartcontractkit/chainlink/core/chains"
+	"github.com/smartcontractkit/chainlink/core/chains/evm/client"
+	v2 "github.com/smartcontractkit/chainlink/core/chains/evm/config/v2"
 	httypes "github.com/smartcontractkit/chainlink/core/chains/evm/headtracker/types"
 	"github.com/smartcontractkit/chainlink/core/chains/evm/log"
 	"github.com/smartcontractkit/chainlink/core/chains/evm/logpoller"
@@ -20,6 +22,7 @@ import (
 	"github.com/smartcontractkit/chainlink/core/chains/evm/types"
 	evmtypes "github.com/smartcontractkit/chainlink/core/chains/evm/types"
 	"github.com/smartcontractkit/chainlink/core/config"
+	cfgv2 "github.com/smartcontractkit/chainlink/core/config/v2"
 	"github.com/smartcontractkit/chainlink/core/logger"
 	"github.com/smartcontractkit/chainlink/core/services"
 	"github.com/smartcontractkit/chainlink/core/services/keystore"
@@ -38,21 +41,26 @@ type ChainConfigUpdater func(*types.ChainCfg) error
 type ChainSet interface {
 	services.ServiceCtx
 	Get(id *big.Int) (Chain, error)
+
 	Show(id utils.Big) (types.DBChain, error)
+
+	Default() (Chain, error)
+	Chains() []Chain
+	ChainCount() int
+
+	ORM() types.ORM
+
 	Add(ctx context.Context, id utils.Big, config *types.ChainCfg) (types.DBChain, error)
 	Remove(id utils.Big) error
-	Default() (Chain, error)
-	Configure(ctx context.Context, id utils.Big, enabled bool, config *types.ChainCfg) (types.DBChain, error)
-	UpdateConfig(id *big.Int, updaters ...ChainConfigUpdater) error
-	Chains() []Chain
 	Index(offset, limit int) ([]types.DBChain, int, error)
-	ChainCount() int
-	ORM() types.ORM
-	// GetNode et al retrieves Nodes from the ORM and adds additional state info
-	GetNode(ctx context.Context, id int32) (node types.Node, err error)
+	UpdateConfig(id *big.Int, updaters ...ChainConfigUpdater) error
+	Configure(ctx context.Context, id utils.Big, enabled bool, config *types.ChainCfg) (types.DBChain, error)
+
+	// GetNodes et al retrieves Nodes from the ORM and adds additional state info
 	GetNodes(ctx context.Context, offset, limit int) (nodes []types.Node, count int, err error)
 	GetNodesForChain(ctx context.Context, chainID utils.Big, offset, limit int) (nodes []types.Node, count int, err error)
 	GetNodesByChainIDs(ctx context.Context, chainIDs []utils.Big) (nodes []types.Node, err error)
+
 	CreateNode(ctx context.Context, data types.Node) (types.Node, error)
 	DeleteNode(ctx context.Context, id int32) error
 }
@@ -63,8 +71,9 @@ type chainSet struct {
 	startedChains []Chain
 	chainsMu      sync.RWMutex
 	logger        logger.Logger
-	orm           types.ORM
 	opts          ChainSetOpts
+
+	immutable bool // toml config is immutable
 }
 
 func (cll *chainSet) Start(ctx context.Context) error {
@@ -86,7 +95,11 @@ func (cll *chainSet) Start(ctx context.Context) error {
 	for i, c := range cll.startedChains {
 		evmChainIDs[i] = c.ID()
 	}
-	cll.logger.Infow(fmt.Sprintf("EVM: Started %d/%d chains, default chain ID is %s", len(cll.startedChains), len(cll.Chains()), cll.defaultID.String()), "startedEvmChainIDs", evmChainIDs)
+	defChainID := "nil"
+	if cll.defaultID != nil {
+		defChainID = cll.defaultID.String()
+	}
+	cll.logger.Infow(fmt.Sprintf("EVM: Started %d/%d chains, default chain ID is %s", len(cll.startedChains), len(cll.Chains()), defChainID), "startedEvmChainIDs", evmChainIDs)
 	return nil
 }
 func (cll *chainSet) Close() (err error) {
@@ -111,6 +124,10 @@ func (cll *chainSet) Ready() (err error) {
 
 func (cll *chainSet) Get(id *big.Int) (Chain, error) {
 	if id == nil {
+		if cll.defaultID == nil {
+			cll.logger.Debug("Chain ID not specified, and default is nil")
+			return nil, errors.New("chain ID not specified, and default is nil")
+		}
 		cll.logger.Debugf("Chain ID not specified, using default: %s", cll.defaultID.String())
 		return cll.Default()
 	}
@@ -124,11 +141,11 @@ func (cll *chainSet) Get(id *big.Int) (Chain, error) {
 }
 
 func (cll *chainSet) Show(id utils.Big) (types.DBChain, error) {
-	return cll.orm.Chain(id)
+	return cll.opts.ORM.Chain(id)
 }
 
 func (cll *chainSet) Index(offset, limit int) ([]types.DBChain, int, error) {
-	return cll.orm.Chains(offset, limit)
+	return cll.opts.ORM.Chains(offset, limit)
 }
 
 func (cll *chainSet) Default() (Chain, error) {
@@ -150,13 +167,13 @@ func (cll *chainSet) Default() (Chain, error) {
 // Requires a lock on chainsMu
 func (cll *chainSet) initializeChain(ctx context.Context, dbchain *types.DBChain) error {
 	// preload nodes
-	nodes, _, err := cll.orm.NodesForChain(dbchain.ID, 0, math.MaxInt)
+	nodes, _, err := cll.opts.ORM.NodesForChain(dbchain.ID, 0, math.MaxInt)
 	if err != nil {
 		return err
 	}
 
 	cid := dbchain.ID.String()
-	chain, err := newChain(ctx, *dbchain, nodes, cll.opts)
+	chain, err := newDBChain(ctx, *dbchain, nodes, cll.opts)
 	if err != nil {
 		return errors.Wrapf(err, "initializeChain: failed to instantiate chain %s", dbchain.ID.String())
 	}
@@ -168,6 +185,9 @@ func (cll *chainSet) initializeChain(ctx context.Context, dbchain *types.DBChain
 }
 
 func (cll *chainSet) Add(ctx context.Context, id utils.Big, config *types.ChainCfg) (types.DBChain, error) {
+	if cll.immutable {
+		return types.DBChain{}, cfgv2.ErrUnsupported
+	}
 	cll.chainsMu.Lock()
 	defer cll.chainsMu.Unlock()
 
@@ -176,7 +196,7 @@ func (cll *chainSet) Add(ctx context.Context, id utils.Big, config *types.ChainC
 		return types.DBChain{}, errors.Errorf("chain already exists with id %s", id.String())
 	}
 
-	dbchain, err := cll.orm.CreateChain(id, config)
+	dbchain, err := cll.opts.ORM.CreateChain(id, config)
 	if err != nil {
 		return types.DBChain{}, err
 	}
@@ -184,10 +204,13 @@ func (cll *chainSet) Add(ctx context.Context, id utils.Big, config *types.ChainC
 }
 
 func (cll *chainSet) Remove(id utils.Big) error {
+	if cll.immutable {
+		return cfgv2.ErrUnsupported
+	}
 	cll.chainsMu.Lock()
 	defer cll.chainsMu.Unlock()
 
-	if err := cll.orm.DeleteChain(id); err != nil {
+	if err := cll.opts.ORM.DeleteChain(id); err != nil {
 		return err
 	}
 
@@ -202,11 +225,14 @@ func (cll *chainSet) Remove(id utils.Big) error {
 }
 
 func (cll *chainSet) Configure(ctx context.Context, id utils.Big, enabled bool, config *types.ChainCfg) (types.DBChain, error) {
+	if cll.immutable {
+		return types.DBChain{}, cfgv2.ErrUnsupported
+	}
 	cll.chainsMu.Lock()
 	defer cll.chainsMu.Unlock()
 
 	// Update configuration stored in the database
-	dbchain, err := cll.orm.UpdateChain(id, enabled, config)
+	dbchain, err := cll.opts.ORM.UpdateChain(id, enabled, config)
 	if err != nil {
 		return types.DBChain{}, err
 	}
@@ -236,8 +262,11 @@ func (cll *chainSet) Configure(ctx context.Context, id utils.Big, enabled bool, 
 }
 
 func (cll *chainSet) UpdateConfig(id *big.Int, updaters ...ChainConfigUpdater) error {
+	if cll.immutable {
+		return cfgv2.ErrUnsupported
+	}
 	bid := utils.NewBig(id)
-	dbchain, err := cll.orm.Chain(*bid)
+	dbchain, err := cll.opts.ORM.Chain(*bid)
 	if err != nil {
 		return err
 	}
@@ -257,7 +286,7 @@ func (cll *chainSet) UpdateConfig(id *big.Int, updaters ...ChainConfigUpdater) e
 		}
 	}
 
-	_, err = cll.orm.UpdateChain(*bid, dbchain.Enabled, &updatedConfig)
+	_, err = cll.opts.ORM.UpdateChain(*bid, dbchain.Enabled, &updatedConfig)
 	if err == nil {
 		chain.Config().Configure(updatedConfig)
 	}
@@ -281,21 +310,11 @@ func (cll *chainSet) ChainCount() int {
 }
 
 func (cll *chainSet) ORM() types.ORM {
-	return cll.orm
-}
-
-func (cll *chainSet) GetNode(ctx context.Context, id int32) (n evmtypes.Node, err error) {
-	n, err = cll.orm.Node(id, pg.WithParentCtx(ctx))
-	if err != nil {
-		err = errors.Wrap(err, "GetNode failed to load node from DB")
-		return
-	}
-	cll.addStateToNode(&n)
-	return
+	return cll.opts.ORM
 }
 
 func (cll *chainSet) GetNodes(ctx context.Context, offset, limit int) (nodes []evmtypes.Node, count int, err error) {
-	nodes, count, err = cll.orm.Nodes(offset, limit, pg.WithParentCtx(ctx))
+	nodes, count, err = cll.opts.ORM.Nodes(offset, limit, pg.WithParentCtx(ctx))
 	if err != nil {
 		err = errors.Wrap(err, "GetNodes failed to load nodes from DB")
 		return
@@ -307,7 +326,7 @@ func (cll *chainSet) GetNodes(ctx context.Context, offset, limit int) (nodes []e
 }
 
 func (cll *chainSet) GetNodesForChain(ctx context.Context, chainID utils.Big, offset, limit int) (nodes []evmtypes.Node, count int, err error) {
-	nodes, count, err = cll.orm.NodesForChain(chainID, offset, limit, pg.WithParentCtx(ctx))
+	nodes, count, err = cll.opts.ORM.NodesForChain(chainID, offset, limit, pg.WithParentCtx(ctx))
 	if err != nil {
 		err = errors.Wrap(err, "GetNodesForChain failed to load nodes from DB")
 		return
@@ -319,7 +338,7 @@ func (cll *chainSet) GetNodesForChain(ctx context.Context, chainID utils.Big, of
 }
 
 func (cll *chainSet) GetNodesByChainIDs(ctx context.Context, chainIDs []utils.Big) (nodes []evmtypes.Node, err error) {
-	nodes, err = cll.orm.GetNodesByChainIDs(chainIDs, pg.WithParentCtx(ctx))
+	nodes, err = cll.opts.ORM.GetNodesByChainIDs(chainIDs, pg.WithParentCtx(ctx))
 	if err != nil {
 		err = errors.Wrap(err, "GetNodesForChain failed to load nodes from DB")
 		return
@@ -331,10 +350,16 @@ func (cll *chainSet) GetNodesByChainIDs(ctx context.Context, chainIDs []utils.Bi
 }
 
 func (cll *chainSet) CreateNode(ctx context.Context, data types.Node) (types.Node, error) {
+	if cll.immutable {
+		return types.Node{}, cfgv2.ErrUnsupported
+	}
 	return cll.opts.ORM.CreateNode(data, pg.WithParentCtx(ctx))
 }
 
 func (cll *chainSet) DeleteNode(ctx context.Context, id int32) error {
+	if cll.immutable {
+		return cfgv2.ErrUnsupported
+	}
 	return cll.opts.ORM.DeleteNode(id, pg.WithParentCtx(ctx))
 }
 
@@ -370,17 +395,21 @@ type ChainSetOpts struct {
 	ORM              types.ORM
 
 	// Gen-functions are useful for dependency injection by tests
-	GenEthClient      func(types.DBChain) evmclient.Client
-	GenLogBroadcaster func(types.DBChain) log.Broadcaster
-	GenLogPoller      func(types.DBChain) logpoller.LogPoller
-	GenHeadTracker    func(types.DBChain, httypes.HeadBroadcaster) httypes.HeadTracker
-	GenTxManager      func(types.DBChain) txmgr.TxManager
+	GenEthClient      func() client.Client
+	GenLogBroadcaster func() log.Broadcaster
+	GenLogPoller      func() logpoller.LogPoller
+	GenHeadTracker    func(httypes.HeadBroadcaster) httypes.HeadTracker
+	GenTxManager      func() txmgr.TxManager
 }
 
 func LoadChainSet(ctx context.Context, opts ChainSetOpts) (ChainSet, error) {
 	if err := checkOpts(&opts); err != nil {
 		return nil, err
 	}
+	if h, ok := opts.Config.(v2.HasEVMConfigs); ok {
+		return NewTOMLChainSet(ctx, opts, h.EVMConfigs())
+	}
+
 	chains, err := opts.ORM.EnabledChains()
 	if err != nil {
 		return nil, errors.Wrap(err, "error loading chains")
@@ -394,10 +423,11 @@ func LoadChainSet(ctx context.Context, opts ChainSetOpts) (ChainSet, error) {
 		id := n.EVMChainID.String()
 		nodes[id] = append(nodes[id], n)
 	}
-	return NewChainSet(ctx, opts, chains, nodes)
+	return NewDBChainSet(ctx, opts, chains, nodes)
 }
 
-func NewChainSet(ctx context.Context, opts ChainSetOpts, dbchains []types.DBChain, nodes map[string][]types.Node) (ChainSet, error) {
+// NewDBChainSet returns a new ChainSet from legacy configuration.
+func NewDBChainSet(ctx context.Context, opts ChainSetOpts, dbchains []types.DBChain, nodes map[string][]types.Node) (ChainSet, error) {
 	if err := checkOpts(&opts); err != nil {
 		return nil, err
 	}
@@ -410,11 +440,12 @@ func NewChainSet(ctx context.Context, opts ChainSetOpts, dbchains []types.DBChai
 		}
 	}
 	var err error
-	cll := &chainSet{defaultChainID, make(map[string]*chain), make([]Chain, 0), sync.RWMutex{}, opts.Logger.Named("ChainSet"), opts.ORM, opts}
+	cll := newChainSet(opts)
+	cll.defaultID = defaultChainID
 	for i := range dbchains {
 		cid := dbchains[i].ID.String()
 		cll.logger.Infow(fmt.Sprintf("Loading chain %s", cid), "evmChainID", cid)
-		chain, err2 := newChain(ctx, dbchains[i], nodes[cid], opts)
+		chain, err2 := newDBChain(ctx, dbchains[i], nodes[cid], opts)
 		if err2 != nil {
 			err = multierr.Combine(err, err2)
 			continue
@@ -427,6 +458,40 @@ func NewChainSet(ctx context.Context, opts ChainSetOpts, dbchains []types.DBChai
 	return cll, err
 }
 
+// NewTOMLChainSet returns a new ChainSet from TOML configuration.
+func NewTOMLChainSet(ctx context.Context, opts ChainSetOpts, chains []*v2.EVMConfig) (ChainSet, error) {
+	if err := checkOpts(&opts); err != nil {
+		return nil, err
+	}
+	opts.Logger = opts.Logger.Named("EVM")
+	var err error
+	cll := newChainSet(opts)
+	cll.immutable = true
+	for i := range chains {
+		cid := chains[i].ChainID.String()
+		cll.logger.Infow(fmt.Sprintf("Loading chain %s", cid), "evmChainID", cid)
+		chain, err2 := newTOMLChain(ctx, chains[i], opts)
+		if err2 != nil {
+			err = multierr.Combine(err, err2)
+			continue
+		}
+		if _, exists := cll.chains[cid]; exists {
+			return nil, errors.Errorf("duplicate chain with ID %s", cid)
+		}
+		cll.chains[cid] = chain
+	}
+	return cll, err
+}
+
+func newChainSet(opts ChainSetOpts) *chainSet {
+	return &chainSet{
+		chains:        make(map[string]*chain),
+		startedChains: make([]Chain, 0),
+		logger:        opts.Logger.Named("ChainSet"),
+		opts:          opts,
+	}
+}
+
 func checkOpts(opts *ChainSetOpts) error {
 	if opts.Logger == nil {
 		return errors.New("logger must be non-nil")
@@ -434,7 +499,14 @@ func checkOpts(opts *ChainSetOpts) error {
 	if opts.Config == nil {
 		return errors.New("config must be non-nil")
 	}
-	if opts.ORM == nil {
+
+	if tomlConfig, ok := opts.Config.(v2.HasEVMConfigs); ok {
+		if opts.ORM != nil {
+			return errors.New("cannot pre-set ORM with TOML config")
+		}
+		opts.ORM = chains.NewORMImmut[utils.Big, *types.ChainCfg, types.Node](tomlConfig.EVMConfigs())
+	} else if opts.ORM == nil {
+		// legacy config only
 		opts.ORM = NewORM(opts.DB, opts.Logger, opts.Config)
 	}
 	return nil
